[update] Write notification to stderr
Upon opening the browser for the user, lets write the notification
for the user to stderr.

We'll use fmt for this instead of logrus because logrus only writes
to stderr when you log with a level of either Warn, Error, Fatal, Panic
and neither of those seemed appropriate for logging a simple notification.