crohacz/print-browser-opening-warning (#58)
* [update] Warn user before opening browser

The browser opens pretty suddenly whenever aws-oidc need to
authenticate with okta. This change gives the user a warning message
in the console before opening the browser and also sleeps for 2 
seconds in order to give the user a chance to digest the message.

We'll use fmt for this instead of logrus because logrus only writes
to stderr when you log with a level of either Warn, Error, Fatal, Panic
and neither of those seemed appropriate for logging a simple notification.