[fix] crohacz/informative-browser-msg (#59)
This is a small attempt at trying to make the browser message more informative after a user authenticates with OIDC. 

I first attempted to use a HTML template in order to make the browser page a lot more friendly, but I struggled to get the static HTML to render. I settled for just inline HTML as a string which is very simplistic.